Fix no-std feature
